debug: 가격 타입 수정

+++ app/src/main/java/com/devik/homebarorder/data/model/CartMenuItem.kt
@@ -3,6 +3,6 @@ package com.devik.homebarorder.data.model
 data class CartMenuItem(
     val menuId: Int,
     val menuName: String,
-    val menuPrice: Int,
+    val menuPrice: Long,
     val menuCount: Int,
 )
